[CHORE](foundation-api): Split mullet_proxy by concern
Replace the tests-vs-code split (01cb787) with a split by concern:
each submodule owns its own `#[cfg(test)] mod tests` inline so
tests live next to the code they exercise.

    routes/mullet_proxy/
        mod.rs      -- module declarations + pub(crate) use handler::ask
        error.rs    -- MulletProxyError + ChromaError impl + its test
        merge.rs    -- merge_user helper + its six unit tests
        handler.rs  -- the axum `ask` handler + the FakeAuth /
                       UnauthorizedAuth / stub-mullet fixtures + the
                       five end-to-end integration tests
